changed data type from int to float and migrated again as div 0 was breaking the loan methods

+++ db/migrate/02_create_loans.rb
@@ -2,10 +2,10 @@ class CreateLoans < ActiveRecord::Migration
   def change
     create_table :loans do |t|
       t.belongs_to :user
-      t.integer :loan_face_value
-      t.integer :loan_present_value
-      t.integer :loan_term
-      t.integer :annual_rate
+      t.float :loan_face_value
+      t.float :loan_present_value
+      t.float :loan_term
+      t.float :annual_rate
       t.timestamps
     end
   end
